Start implementing the full environment stack.
Shiro's assumption about the current subject being bound to the current
thread just does not seem to map well to the way clojure handles
multi-threading in general. It smells even fishier when you mix in
something like Stuart Sierra's Components.

Start laying the foundation to have the Security Manager be one Component
that associates a Subject Component with some sort of Request Component
that can be passed around to everything that might need it.