Get it to build
Mainly by tweaking basic project settings.

Also eliminated the dependency on midje.

My unit tests (such as they are) fail, but that's because basic system
start is throwing a not-implemented exception.

The main point for tonight's exercise was to get the basic
infrastructure into place so I can start working on this again.

Someday.